Move Core.FileNameConstants to Projects.ProjectConstants
