+++ Celbridge/CoreServices/Celbridge.Projects/Services/Project.cs
@@ -103,7 +103,7 @@ public class Project : IDisposable, IProject
             var projectPath = Path.GetDirectoryName(projectFilePath);
             Guard.IsNotNull(projectPath);
 
-            var projectDataFolderPath = Path.Combine(projectPath, FileNameConstants.ProjectDataFolder);
+            var projectDataFolderPath = Path.Combine(projectPath, ProjectConstants.ProjectDataFolder);
 
             if (!Directory.Exists(projectDataFolderPath))
             {
@@ -132,7 +132,7 @@ public class Project : IDisposable, IProject
         _projectFolderPath = Path.GetDirectoryName(projectFilePath)!;
         Guard.IsNotNullOrWhiteSpace(ProjectFolderPath);
 
-        _projectDataFolderPath = Path.Combine(ProjectFolderPath, FileNameConstants.ProjectDataFolder);
+        _projectDataFolderPath = Path.Combine(ProjectFolderPath, ProjectConstants.ProjectDataFolder);
     }
 
     private bool _disposed = false;
