+++ Celbridge/CoreServices/Celbridge.Projects/Services/ProjectService.cs
@@ -49,7 +49,7 @@ public class ProjectService : IProjectService
         }
 
         var extension = Path.GetExtension(projectName);
-        if (extension != FileNameConstants.ProjectFileExtension)
+        if (extension != ProjectConstants.ProjectFileExtension)
         {
             return Result.Fail($"Project file extension is not valid: '{projectName}'");
         }
